requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
pydantic-settings
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==4.0.1
slowapi
redis
